requirements.txt
numpy>=1.24.0
pandas>=1.5.0
scikit-learn>=1.0.2
scipy>=1.7.3
scikit-surprise>=1.1.1
flask>=2.0.2
flask-cors>=3.0.10
python-dotenv>=0.19.2
mlflow>=2.8.0
fastapi>=0.104.1
uvicorn>=0.24.0
sqlalchemy>=2.0.23
psycopg2-binary>=2.9.9
alembic>=1.12.1
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
pydantic>=2.5.0
